# backend/src/projects/dto/gateway-configuration.dto.ts
import { IsNotEmpty } from "class-validator";
import { ApiProperty } from "@nestjs/swagger";

export class GatewayConfigurationDto {
    @ApiProperty()
    @IsNotEmpty()
    port: number;

    @ApiProperty()
    @IsNotEmpty()
    address: string;
}